+++ web/src/Tests/UI/Header.cy.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Header from "../../components/Header";
+import Header from "../../components/Header/Header";
 
 describe("<Header />", () => {
   beforeEach(() => {
